タスクのドラッグ&ドロップとルーティング修正

+++ app/controllers/admin/users_controller.rb
@@ -11,7 +11,9 @@ class Admin::UsersController < ApplicationController
 
   def destroy
     @user = find_user_by_id
+    # pp @user
     if !@user.admin?
+      pp @user
       destroy_task
       @user.destroy
       flash[:notice] = "ユーザとタスク削除完了"
@@ -52,6 +54,8 @@ class Admin::UsersController < ApplicationController
 
   def destroy_task
     task = Task.find_by_user_id(find_user_by_id)
-    task.destroy
+    if task.present?
+      task.destroy
+    end
   end
 end
